+++ src/main/java/com/gruzini/tennistico/models/dto/SetDTO.java
@@ -1,4 +1,4 @@
-package com.gruzini.tennistico.models.score;
+package com.gruzini.tennistico.models.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
