+++ src/main/java/com/gruzini/tennistico/services/InputScoreService.java
@@ -1,26 +1,31 @@
 package com.gruzini.tennistico.services;
 
 import com.gruzini.tennistico.domain.Match;
-import com.gruzini.tennistico.domain.Player;
-import com.gruzini.tennistico.models.score.WinDecider;
+import com.gruzini.tennistico.domain.enums.MatchStatus;
+import com.gruzini.tennistico.services.score.WinDecider;
 import org.springframework.stereotype.Service;
 
 @Service
 public class InputScoreService {
 
+    private final MatchService matchService;
     private final RankingService rankingService;
     private final WinDecider winDecider;
 
-    public InputScoreService(RankingService rankingService, WinDecider winDecider) {
+    public InputScoreService(MatchService matchService, RankingService rankingService, WinDecider winDecider) {
+        this.matchService = matchService;
         this.rankingService = rankingService;
         this.winDecider = winDecider;
     }
 
-    public void inputScore(final Match match) {
-        final Player winner = winDecider.decideWinner(match);
-        final Player loser = winDecider.decideLoser(match);
+    public void inputScore(final Long matchId, final String score) {
+        final Match match = matchService.getById(matchId);
 
-        rankingService.updateWinner(winner);
-        rankingService.updateLoser(loser);
+        matchService.updateMatchScore(match, score);
+
+        rankingService.updateWinner(winDecider.decideWinner(match));
+        rankingService.updateLoser(winDecider.decideLoser(match));
+
+        matchService.updateMatchStatus(match, MatchStatus.SCORE_TO_BE_CONFIRMED);
     }
 }
