+++ src/test/java/com/gruzini/tennistico/services/RankingServiceTest.java
@@ -2,7 +2,7 @@ package com.gruzini.tennistico.services;
 
 import com.gruzini.tennistico.domain.Player;
 import com.gruzini.tennistico.domain.Ranking;
-import com.gruzini.tennistico.models.score.RankingPointCounter;
+import com.gruzini.tennistico.services.score.RankingPointCounter;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
